iprt/RTDirCreateUniqueNumbered: build fix
svn:sync-xref-src-repo-rev: r71049

+++ src/VBox/Runtime/generic/RTDirCreateUniqueNumbered-generic.cpp
@@ -1,4 +1,4 @@
-/* $Id: RTDirCreateUniqueNumbered-generic.cpp 36600 2011-04-06 20:06:31Z knut.osmundsen@oracle.com $ */
+/* $Id: RTDirCreateUniqueNumbered-generic.cpp 36604 2011-04-06 23:02:52Z noreply@oracle.com $ */
 /** @file
  * IPRT - RTDirCreateUniqueNumbered, generic implementation.
  */
@@ -83,7 +83,7 @@ RTDECL(int) RTDirCreateUniqueNumbered(char *pszPath, size_t cbSize, RTFMODE fMod
     {
         /* Format the number with leading zero's. */
         ssize_t rc2 = RTStrFormatU32(pszEnd, cbLeft, i, 10, cchDigits, 0, RTSTR_F_WIDTH | RTSTR_F_ZEROPAD);
-        if (RT_FAILURE(rc2))
+        if (RT_FAILURE((int) rc2))
         {
             *pszPath = '\0';
             return (int)rc2;
